fix(wave-autoscale): file name change and add test case
